Clean up RobotId and its usage
